some changes

+++ database/migrations/2021_07_14_100917_create_buch__kopfs_table.php
@@ -17,7 +17,7 @@ class CreateBuchKopfsTable extends Migration
             $table->id();
             $table->char('name', 100);
             $table->char('spedition', 100);
-            $table->integer("buchungsnummer");
+            $table->bigInteger("buchungsnummer");
             $table->timestamps();
         });
     }
